+++ pytorch_lightning/trainer/connectors/accelerator_connector.py
@@ -803,13 +803,6 @@ class AcceleratorConnector:
         return 0
 
     @property
-    def tpu_id(self) -> Optional[int]:
-        if isinstance(self.accelerator, TPUAccelerator):
-            if isinstance(self._tpu_cores, list):
-                return self._tpu_cores[0]
-        return None
-
-    @property
     def num_ipus(self) -> int:
         if isinstance(self.accelerator, IPUAccelerator):
             return self.devices